+++ packages/app-root/src/components/PageHeader.js
@@ -1,11 +1,8 @@
 'use client'
-import { ZooHeader } from '@zooniverse/react-components'
-import { useContext } from 'react'
-
-import {
-  useUnreadMessages,
-  useUnreadNotifications
-} from '../hooks'
+import { AuthModal, ZooHeader } from '@zooniverse/react-components'
+import { useUnreadMessages, useUnreadNotifications } from '@zooniverse/react-components/hooks'
+import auth from 'panoptes-client/lib/auth'
+import { useContext, useState } from 'react'
 
 import { PanoptesAuthContext } from '../contexts'
 
@@ -13,11 +10,36 @@ export default function PageHeader() {
   const { adminMode, user } = useContext(PanoptesAuthContext)
   const { data: unreadMessages }= useUnreadMessages(user)
   const { data: unreadNotifications }= useUnreadNotifications(user)
+  const [activeIndex, setActiveIndex] = useState(-1)
+
+  function openRegisterModal() {
+    setActiveIndex(1)
+  }
+
+  function openSignInModal() {
+    setActiveIndex(0)
+  }
+
+  function onSignOut() {
+    auth.signOut()
+  }
+
+  function closeAuthModal() {
+    setActiveIndex(-1)
+  }
 
   return (
     <header aria-label='Zooniverse site header'>
+      <AuthModal
+        activeIndex={activeIndex}
+        closeModal={closeAuthModal}
+        onActive={setActiveIndex}
+      />
       <ZooHeader
         isAdmin={adminMode}
+        register={openRegisterModal}
+        signIn={openSignInModal}
+        signOut={onSignOut}
         unreadMessages={unreadMessages}
         unreadNotifications={unreadNotifications}
         user={user}