+++ packages/lib-react-components/src/hooks/index.js
@@ -1,2 +1,5 @@
 export { default as useJSONData } from './useJSONData'
+export { default as usePanoptesUser } from './usePanoptesUser.js'
+export { default as useUnreadMessages } from './useUnreadMessages.js'
+export { default as useUnreadNotifications } from './useUnreadNotifications.js'
 export { default as useProgressiveImage } from './useProgressiveImage'